Move the test to the more appropriate module
